Menu UI fidelity wave 2: retail-pinned Status/Equip/Options/top-level + real XP curve (#295)
* Equip screen: retail four-window layout traced from FUN_801D21C0/FUN_801D2094

Replace the invented full-screen equip list with the retail window set
(tab 2 + party 21 + item-list 23 + main 22, rects from the boot-parsed
menu-window descriptor table): "Best Equipment" header, 0xE-pitch slot
rows with the row-8 ICO pictogram column and hand cursors from the
system-UI atlas, stat-compare block at the traced main-window offsets,
candidates in the id-23 list window. Slot navigation/equip flow stays in
engine-core; docs + site fragment gain the traced Equip-screen section.


* Retail options screen: traced row/popup model replaces the invented Config rows

Trace (menu overlay PROT 0899): the id-48 window renderer FUN_801DCEF0 wraps
row renderer FUN_801D2910, driven by the display-layout table 0x801E4404
([row_id, advance] pairs; 14/20 px), the 8-byte row-descriptor list 0x801E44B8
(config-word ptr, value count, label ink, row id, string index) and the string
pointer table 0x801E442C (value = strings[label + value + 1]). Input SM
FUN_801DA9F8 skips valueless rows, opens the value popup (window descriptor
id 47 - x/w static (170,128), y/h stamped per open, flip above past y=0xB0)
and commits the choice straight into the 0x800845xx/0x800846xx config words;
FUN_801D2B44 renders the popup at a 13-px pitch. Hidden "Battle Voices" row
exists in the descriptors but not the layout.

Engine: engine-core::options rewritten around the retail row set (Battle
Camera/Select Attack/Command, Field Move/HP Display, Sound, Dual Shock
Battles/Events/Encounters) with the Browsing->Editing popup SM and
options_popup_content_rect; engine-only knobs (BGM/SFX volume, message speed)
stay in the new legaia-options.toml round-trip but leave the menu rows; the
footer hint is gone (retail has none). options_draws_for now draws the pinned
geometry (labels x+8 white/teal, values x+140 gold, capture-sampled inks) plus
the popup; the shell chromes the popup window and Sound Stereo/Monaural drives
a new AudioOut::set_mono downmix. Docs: field-menu.md options-screen section +
site fragment + functions.md entry.


* Real retail XP curve shipped as derived constants; record XP fields realigned

The engine's disc-less XP fallback (engine-core retail_xp_table +
legaia_save::RETAIL_XP_CUMULATIVE) still carried the refuted GTE sin-LUT
slice (50, 56, 62, ...), and the retail New Game Status capture shows
"Next Level 121" - the real L2 threshold. The DAT_80076AF4 delta table
turns out to be the closed form delta(n) = n^2/4 + 1, so the whole curve
is derivable arithmetic (FUN_801E9504: sum scaled by 9_999_999/0x140FE
below level 0x11, else x 0x79): both tables now build the real base
curve (121, 365, 730, 1338, 2190, ..., 9_646_483) from the formula - no
Sony table bytes committed. The boot-time disc parse stays installed and
a disc-gated assert pins derived == disc-parsed.

Empirically validated by mednafen RAM sampling across the save-state
library: every (level, record +0x4) pair L1..L37 matches, including the
Noa/Gala sin-divisor corrections (New Game 121/102/140); L99 carries 0.

CharacterRecord accessors realigned to the retail record: cumulative XP
is u32 at +0x0 (was u16 at +0x4 - the low half of the threshold), new
next_level_xp accessor at +0x4; the status renderer FUN_801D33D8 draws
+0x0/+0x4 verbatim (Next Level = the cumulative threshold, not the
remaining difference), so the engine status screen now does too. New
Game seeds Experience 0 / Next Level 121; apply_battle_xp re-stamps
+0x0/+0x4/+0x130 after every grant; load_party hydrates the tracker XP +
level from the records. Docs + site fragment updated; schema fixtures
regenerated.


* Status-screen retail fidelity: tab plaque, satellite sprites, retail inks + digit cells, persistent AP

Pixel-faithful pause-menu STATUS screen against the golden
menu_status_town capture, plus the shared retail tab-banner primitive:

- tab_banner_draws (engine-render): the carved brown plaque behind every
  screen's title tab - left cap (208,64,8,20) + tiled 16x20 body
  (192,64) + right cap (216,64,8,20), CLUT row 12, at (WX-8, WY-4)
  (RAM GPU-prim scan over the menu_status_town state). Class-2 tab
  windows draw NO gold 9-slice frame; Status/Options screens wired,
  Equip inherits the same path when its window set lands.
- Satellite sprites (status_satellite_icon_sprites_for): party-list
  pointing-hand cursor (FUN_801D2094: names at WX+6, pitch 0xe, hand =
  FUN_8002b994 sprite-table kind 0 at WX-0xc), Condition-pager solid
  triangles (FUN_801D30A4: kinds 2/3, UV (168,8)/(168,40), at
  WX-0x10/WX+0x3A, WY-2), summary-window LV icon + 2-digit level +
  ATR element icon (FUN_801D31EC; the 0xCE token resolves through the
  0x80074050 aux table to ICO codes 0x94/0x96/0x95 - 28x12 sprites in
  the system-UI extension-strip TIM at PROT.DAT[0x10178], decoded with
  the VRAM-row-500 CLUT block of the sibling TIM at PROT.DAT[0x10028]).
- Retail number cells + inks: FUN_80034b78 fields lay one glyph per
  8-px cell ("180/ 180 ( 180)" spacing); parenthesised base/growth
  groups in the retail teal (66,222,222), body text in the CLUT-7
  white (206,206,206) - linear-space constants so the sRGB surface
  presents the exact retail bytes. Stat-grid parens pinned at
  +0x40/+0x60 (left) and +0xb4/+0xd4 (right).
- Window frame margin corrected to the prim-scanned content-8 on all
  four sides (was -6/-2/+10); every menu window frame now byte-matches
  the retail border rows.
- Status AP gauge feeds the persistent char-record +0x10E AP (0 on a
  fresh party; new-game seed zeroes it) instead of the battle ApGauge
  base; battle gauge behavior unchanged.
- Footer hint line removed from the status screen (retail has none).
- docs/subsystems/field-menu.md + site fragment: tab-banner + satellite
  renderer sections, 0xCE/ICO/NUM/CUR primitive semantics, extension
  strip provenance, frame-margin and ink pins.


* Top-level pause menu retail content + cross-branch seams (hand cursors, shared AP gauge)

Traces FUN_801CFD68 / FUN_801D0148 / FUN_801D030C (top-level command
list, money/play-time box, party info panel) and replaces the invented
engine content: retail row set Items/Magic/Equip/Status/Options/Load/
Save at (WX+0x14, WY+n*0xe), money pictogram + 8-digit amount +
H:MM:SS clock with zero-padded fields, and per-member party blocks
(stride 0x3e) with LV/HP/MP label sprites and the kind-0x31 AP gauge
fed from persistent AP. FieldMenuRow restructured to the retail rows
(Load = SaveSelect in Load mode wired to load_from_slot; the Arts
chain editor stays a directly-constructible engine extension).

Cross-branch integration seams: the options screen selected row now
draws the retail pointing-hand sprite (options_hand_cursor_sprite)
instead of ">", the AP gauge widget is factored into the shared
ap_gauge_sprites, and the money/time/coin ICO records (0x62/0x63/0x66,
label CLUT row) join the boot-built system-UI atlas.

Docs: field-menu.md + site fragment gain the top-level section; the
engine-port paragraphs drop the stale "top-level untraced" claim.


---------

+++ crates/engine-core/tests/end_to_end_gameplay_loop.rs
@@ -92,7 +92,7 @@ fn synthetic_party() -> Party {
             sp_max,
         });
         rec.set_level(level);
-        rec.set_cumulative_xp(xp as u16);
+        rec.set_cumulative_xp(xp);
         rec.set_stat_cap(0x3E7);
     }
     p
